requirements.txt
dnspython==2.0.0
ipwhois==1.2.0
python-dotenv==1.0.0